backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
celery
redis
dnspython
aioredis